fixed SLTI and added MULT and DIV back in
multiplication and division currently happens instantly instead of after 6-36
cycles like expected. I could easily use the methods made for the load delay
slot to implement the delay, but mult and div have the additional complication
that if the CPU tries to access the results before the operation is done, the
CPU stalls meaning that this will be a bit more involved.

+++ src/register.rs
@@ -76,9 +76,8 @@ impl Parts for Register{
       0
     }
   }
-  //FIXME: fix signed comparison
   fn signed_compare(&self, rhs: Register) -> Register {
-    if *self < rhs {
+    if (*self as i32) < (rhs as i32) {
       1
     } else {
       0
